Refactored app folder name to redux
Also updated import of dependents
